Config Module + Temporary Enviro Setting
Core of the Configuration Module which loads the config.js file - sets the paths to ensure they have a single reference point and are standardized throughout the application.

Temporary enviro.js to control the environment setting. Additional temporary setting for this dropped into boot loader as well.

Minor updates to copyrights, readme, etc. to reflect it being 2014.

Signed-off-by: Scott Szarapka <scott@szarapka.com>