First Iteration of S3 Uploading, It's Ugly
This is the first ugly version of S3 uploading, the uploads work, thats about the only good thing so far haha.

Also added GM to the packages list for image manipulations.

Still torn on straight streaming to S3 and bypassing tmp entirely, cause image manipulation might want to take place before the file is 'saved'. Either way it's a simple change to pipe directly to S3 instead of the way files are stored in yield/tmp and then tossed up.

Few minor tweaks and fixes as well, typos, etc.

Signed-off-by: Scott Szarapka <scott@szarapka.com>